Split place_module to separated finders
