+++ .gitignore
@@ -1,4 +1,5 @@
 *.py[cod]
+__pycache__
 .DS_Store
 # C extensions
 *.so
@@ -25,6 +26,7 @@ npm-debug.log
 
 # Unit test / coverage reports
 .coverage
+.pytest_cache
 .tox
 nosetests.xml
 htmlcov
